Update Readme.md file with additional features and naming
The Readme.md file has been updated in this commit which includes the addition of four new features in the list and also changing the name 'PHPSpreadsheet Report Generator' to 'Reptor' under installation/usage section. The newly added features provide more insight into the capabilities of the software. The name change reflects the actual name of the project.