Update count logic in Avg aggregator
The count logic in the Avg aggregator was previously based on the subject value being greater than zero. This commit updates this logic to check whether the string length of the subject is greater than zero instead. Additionally, it removes the use of null coalescing in the sum update to capture the cases where the subject is a valid value of 0.